Module accurate::dot

Algorithms for dot product

Modules

traits
Dot product traits

Structs

Dot2
DotK with two cascaded accumulators

DotConsumer
Adapts a ParallelDotAccumulator into a Consumer

DotFolder
Adapts a DotAccumulator into a Folder

DotK
Calculates a dot product using both product transformation and cascaded accumulators

NaiveDot
Naive floating point dot product

OnlineExactDot
Calculates the dot product using product transformation and OnlineExactSum

Type Definitions

Dot3
DotK with three cascaded accumulators

Dot4
DotK with four cascaded accumulators

Dot5
DotK with five cascaded accumulators

Dot6
DotK with six cascaded accumulators

Dot7
DotK with seven cascaded accumulators

Dot8
DotK with eight cascaded accumulators

Dot9
DotK with nine cascaded accumulators