Module accurate::sum

Algorithms for summation

Modules

traits
Summation traits

Structs

NaiveSum
Naive floating point summation

OnlineExactSum
Calculates a sum using separate accumulators for each possible exponent

Sum2
SumK with two cascaded accumulators

SumConsumer
Adapts a ParallelSumAccumulator into a Consumer

SumFolder
Adapts a SumAccumulator into a Folder

SumK
Calculates a sum using cascaded accumulators for the remainder terms

Functions

i_fast_sum_in_place
Calculates the correctly rounded sum of numbers in a slice

Type Definitions

Sum3
SumK with three cascaded accumulators

Sum4
SumK with four cascaded accumulators

Sum5
SumK with five cascaded accumulators

Sum6
SumK with six cascaded accumulators

Sum7
SumK with seven cascaded accumulators

Sum8
SumK with eight cascaded accumulators

Sum9
SumK with nine cascaded accumulators