accurate

Module traits

Includes all traits of this crate

Traits

DotAccumulator

Accumulates terms of a dot product

DotWithAccumulator

Calculates the dot product of the items of an iterator

FloatFormat

Describes the layout of a floating-point number

HalfUlp

Half a unit in the last place (ULP)

IFastSum

Calculates the correctly rounded sum of numbers in a slice

RawExponent

Extract the raw exponent of a floating-point number

Round3

Correctly rounded sum of three non-overlapping numbers

Split

Split a floating-point number

SumAccumulator

Accumulates terms of a sum

SumWithAccumulator

Sums the items of an iterator

TwoProduct

Product transformation

TwoSum

Sum transformation