Trait accurate::traits::ParallelDotWithAccumulator

pub trait ParallelDotWithAccumulator<F>: ParallelIterator<Item = (F, F)> where
    F: Send, {
    fn parallel_dot_with_accumulator<Acc>(self) -> F
    where
        Acc: ParallelDotAccumulator<F>,
        F: Zero,
    { ... }
}

Calculates the dot product of an iterator, possibly in parallel

Examples

use rayon::prelude::*;

use accurate::traits::*;
use accurate::dot::OnlineExactDot;

let d = vec![(1.0, 1.0), (2.0, 2.0), (3.0, 3.0)]
    .par_iter().map(|&x| x)
    .parallel_dot_with_accumulator::<OnlineExactDot<_>>();
assert_eq!(14.0f64, d);

Provided methods

fn parallel_dot_with_accumulator<Acc>(self) -> F where
Acc: ParallelDotAccumulator<F>,
F: Zero,

Calculate the dot product of an iterator, possibly in parallel

Loading content...