Trait accurate::util::traits::FloatFormat

pub trait FloatFormat {
    fn base() -> u32;

    fn significand_digits() -> u32;

    fn exponent_digits() -> u32;

    fn base_pow_exponent_digits() -> usize { ... }

    fn base_pow_significand_digits_half() -> usize { ... }
}

Describes the layout of a floating-point number

Required methods

fn base() -> u32

The number format's base

fn significand_digits() -> u32

The length of the number format's significand field

fn exponent_digits() -> u32

The length of the number format's exponent field

Loading content...
Provided methods

fn base_pow_exponent_digits() -> usize

The base raised to the power of the exponent`s length

fn base_pow_significand_digits_half() -> usize

The base raised to the power of half the mantissa`s length

Loading content...