Module accurate::traits

Includes all traits of this crate

Traits

DotAccumulator
Accumulates terms of a dot product

DotWithAccumulator
Calculates the dot product of the items of an iterator

FloatFormat
Describes the layout of a floating-point number

HalfUlp
Half a unit in the last place (ULP)

IFastSum
Calculates the correctly rounded sum of numbers in a slice

ParallelDotAccumulator
A DotAccumulator that can be used in parallel computations

ParallelDotWithAccumulator
Calculates the dot product of an iterator, possibly in parallel

ParallelSumAccumulator
A SumAccumulator that can be used in parallel computations

ParallelSumWithAccumulator
Sums the items of an iterator, possibly in parallel

RawExponent
Extract the raw exponent of a floating-point number

Round3
Correctly rounded sum of three non-overlapping numbers

Split
Split a floating-point number

SumAccumulator
Accumulates terms of a sum

SumWithAccumulator
Sums the items of an iterator

TwoProduct
Product transformation

TwoSum
Sum transformation